requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning dependencies (CPU-only versions for smaller images)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.4.1
torchvision==0.19.1
tensorflow-cpu==2.16.1
pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
onnxruntime==1.16.3

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
cachetools==5.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring
prometheus-client==0.19.0

# Logging
python-json-logger==2.0.7
structlog==23.2.0

# Performance Testing
locust==2.17.0
aiohttp==3.9.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
